fix: prevent circular imports (jfc)
Fixes circular imports when declaring the player object. Both
`player.talents` and `player.barFragments` were either classes, or a
<key, class> object pair. In some instances, these classes would
reference the player object, but because they were being instantiated as
the player object was being initialized, it would cause the value of
`player` in these classes to be undefined/non-existent. By declaring the
object first, and then dynamically adding these properties
(`coinUpgrades`, `barFragments`, and `talents`), we can ensure that the
player object is always defined when accessing it.